+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.65.5] 2023-08-17
+
+### Added
+
+- Expected keywords now uses keywords
+
+## [3.65.3-4] 2023-08-08
+
+### Fixed
+
+- Expected keywords now uses keywords
+
 ## [3.65.0-1] 2023-08-02
 
 ### Added